CORTX-26259: addressing review comments
